ADD : 문 열기 실패시 실패 효과음 추가
