+++ Prison Escape/.vscode/extensions.json
@@ -1,0 +1,5 @@
+{
+    "recommendations": [
+      "visualstudiotoolsforunity.vstuc"
+    ]
+}
