--- Prison Escape/Assets/Scripts/Chess/Door.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 068e2aae9bfff754d83b38337e2ef1fe
