--- Prison Escape/Assets/Scripts/Prison/PrisonDoor.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: f4da122e3720dd149966b61cd78d2dae
